third_party/WebKit/LayoutTests/http/tests/security/cross-frame-access-call.html

<html>
<head>
    <script src="/js-test-resources/js-test-pre.js"></script>
    <script src="resources/cross-frame-access.js"></script>
</head>
<body>
<iframe src="http://localhost:8000/security/resources/cross-frame-iframe-for-get-test.html" style=""></iframe>
<pre id="console"></pre>
<script>

window.jsTestIsAsync = true;
window.targetWindow = frames[0];

window.onload = function()
{
    if (window.testRunner)
        testRunner.setCanOpenWindows(true);

    description("\n----- tests for calling methods of another frame using Function.call -----\n");

    // Allowed
    // void focus();
    // void blur();
    // void close();
    // void postMessage(in DOMString message);

    // - Tests for the Window object - 
    // undefined value indicates failure
    shouldThrow("window.setTimeout.call(targetWindow, 'void(0);', 0)", '"SecurityError: Blocked a frame with origin \\"http://127.0.0.1:8000\\" from accessing a cross-origin frame."');
    shouldThrow("window.setInterval.call(targetWindow, 'void(0);', 0)", '"SecurityError: Blocked a frame with origin \\"http://127.0.0.1:8000\\" from accessing a cross-origin frame."');
    shouldBeUndefined("window.getSelection.call(targetWindow)");
    shouldBeUndefined("window.find.call(targetWindow, 'string', false, false, false, false, false, false)");
    shouldBeUndefined("window.confirm.call(targetWindow, 'message')");
    shouldBeUndefined("window.prompt.call(targetWindow, 'message', 'defaultValue')");
    shouldBeUndefined("window.getComputedStyle.call(targetWindow, document.body, '')");
    shouldBeUndefined("window.getMatchedCSSRules.call(targetWindow, document.body, '')");
    shouldBeUndefined("window.openDatabase.call(targetWindow, 'name', '1.0', 'description', 0)");
    shouldBeUndefined("window.atob.call(targetWindow, 'string')");
    shouldBeUndefined("window.btoa.call(targetWindow, 'string')");
    shouldBeUndefined("window.open.call(targetWindow, '')");

    // These always return undefined so we use the error console to detect failure
    shouldBeUndefined("window.addEventListener.call(targetWindow, 'load', null, false);");
    shouldBeUndefined("window.removeEventListener.call(targetWindow, 'load', null, false);");
    shouldBeUndefined("window.clearTimeout.call(targetWindow, 0);");
    shouldBeUndefined("window.clearInterval.call(targetWindow, 0);");
    shouldBeUndefined("window.print.call(targetWindow);");
    shouldBeUndefined("window.stop.call(targetWindow);");
    shouldBeUndefined("window.alert.call(targetWindow, 'message');");
    shouldBeUndefined("window.scrollBy.call(targetWindow, 0, 0);");
    shouldBeUndefined("window.scrollTo.call(targetWindow, 0, 0);");
    shouldBeUndefined("window.scroll.call(targetWindow, 0, 0);");
    shouldBeUndefined("window.moveBy.call(targetWindow, 0, 0);");
    shouldBeUndefined("window.moveTo.call(targetWindow, 0, 0);");
    shouldBeUndefined("window.resizeBy.call(targetWindow, 0, 0);");
    shouldBeUndefined("window.resizeTo.call(targetWindow, 0, 0);");
    shouldBeUndefined("window.showModalDialog.call(targetWindow);");

    // Throws an EvalError and logs to the error console
    shouldBe("window.eval.call(targetWindow, '1+2');", '3');

    // - Tests for the Location object -
    // undefined value indicates failure
    shouldBeUndefined("window.location.toString.call(targetWindow.location)");

    // Work around DRT bug that causes subsequent tests to fail.
    window.stop();
    finishJSTest();
}
</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
